fix exhaustive try promotion flow
