+++ crates/sifr_ir/src/lib.rs
@@ -8,6 +8,7 @@ pub mod cfg;
 pub mod diagnostic_types;
 pub mod flow_graph;
 mod hir_expr;
+mod hir_flow;
 pub mod hir_nodes;
 pub mod lowering_outcome;
 pub mod lowering_result;
@@ -19,6 +20,7 @@ mod type_visit;
 pub use cfg::*;
 pub use diagnostic_types::*;
 pub use flow_graph::*;
+pub use hir_flow::*;
 pub use hir_nodes::*;
 pub use lowering_outcome::LoweringOutcome;
 pub use lowering_result::LoweringResult;
